Fixed some MSVC warnings

+++ src/util.cpp
@@ -1210,8 +1210,8 @@ void ShrinkDebugFile()
     {
         // Restart the file with some of the end
         char pch[200000];
-        fseek(file, -sizeof(pch), SEEK_END);
-        int nBytes = fread(pch, 1, sizeof(pch), file);
+        fseek(file, -((long long)sizeof(pch)), SEEK_END);
+        size_t nBytes = fread(pch, 1, sizeof(pch), file);
         fclose(file);
 
         file = fopen(pathLog.string().c_str(), "w");
